backend/requirements.txt
flask
flask-cors
git+https://github.com/supabase/supabase-py
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
websockets>=11.0.3
google-generativeai

